+++ agrest-engine/src/main/java/io/agrest/meta/LinkMethodType.java
@@ -1,9 +1,0 @@
-package io.agrest.meta;
-
-/**
- * @deprecated since 4.1, as Agrest now integrates with OpenAPI 3 / Swagger.
- */
-@Deprecated
-public enum LinkMethodType {
-    GET, POST, PUT, DELETE
-}