+++ agrest-jaxrs3/src/main/java/io/agrest/jaxrs3/AgJaxrsFeatureBuilder.java
@@ -2,13 +2,11 @@ package io.agrest.jaxrs3;
 
 import io.agrest.DataResponse;
 import io.agrest.EntityUpdate;
-import io.agrest.MetadataResponse;
 import io.agrest.SimpleResponse;
 import io.agrest.jaxrs3.provider.DataResponseWriter;
 import io.agrest.jaxrs3.provider.EntityUpdateCollectionReader;
 import io.agrest.jaxrs3.provider.EntityUpdateReader;
 import io.agrest.jaxrs3.provider.JaxrsAgExceptionMapper;
-import io.agrest.jaxrs3.provider.MetadataResponseWriter;
 import io.agrest.jaxrs3.provider.ResponseStatusDynamicFeature;
 import io.agrest.jaxrs3.provider.SimpleResponseWriter;
 import io.agrest.runtime.AgRuntime;
@@ -59,9 +57,6 @@ public class AgJaxrsFeatureBuilder {
 
         map.put(SimpleResponse.class, SimpleResponseWriter.class);
         map.put(DataResponse.class, DataResponseWriter.class);
-
-        // deprecated writers
-        map.put(MetadataResponse.class, MetadataResponseWriter.class);
     }
 
     public AgJaxrsFeature build() {